SW-1749 Split up AccessionStoreTest
- Create a base class
  - base class inherits DatabaseTest
  - implements RunAsUser
  - supplies helper vals/funs
- Categorize existing tests into new, smaller test classes
