+++ src/test/kotlin/com/terraformation/backend/seedbank/db/accessionStore/AccessionStorePhotoTest.kt
@@ -1,0 +1,37 @@
+package com.terraformation.backend.seedbank.db.accessionStore
+
+import com.terraformation.backend.db.default_schema.tables.pojos.PhotosRow
+import com.terraformation.backend.db.seedbank.AccessionId
+import com.terraformation.backend.db.seedbank.tables.pojos.AccessionPhotosRow
+import com.terraformation.backend.seedbank.model.AccessionModel
+import java.net.URI
+import java.time.Instant
+import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Test
+import org.springframework.http.MediaType
+
+internal class AccessionStorePhotoTest : AccessionStoreTest() {
+  @Test
+  fun `photo filenames are returned`() {
+    val initial = store.create(AccessionModel(facilityId = facilityId))
+    val photosRow =
+        PhotosRow(
+            fileName = "photo.jpg",
+            createdBy = user.userId,
+            createdTime = Instant.now(),
+            contentType = MediaType.IMAGE_JPEG_VALUE,
+            modifiedBy = user.userId,
+            modifiedTime = Instant.now(),
+            size = 123,
+            storageUrl = URI("file:///photo.jpg"),
+        )
+    photosDao.insert(photosRow)
+
+    accessionPhotosDao.insert(
+        AccessionPhotosRow(accessionId = AccessionId(1), photoId = photosRow.id))
+
+    val fetched = store.fetchOneById(initial.id!!)
+
+    Assertions.assertEquals(listOf("photo.jpg"), fetched.photoFilenames)
+  }
+}
